Fix account creation with the registration app
When an account is created from the registration app there is no user
session. This lead to an Error in createUser() as $adminUser is used
while null, but only Exceptions are cought there.

Add a check that throw an Exception when $adminUser is null to
fallback in the code path that handle the case where the admin user is
not from the LDAP directory.

Signed-off-by: Alban Bedel <albeu@free.fr>