Prevent LDAP injection when creating accounts
The LDIF template generation just paste the username and password as
is in the template without any check, thus allowing to inject any LDAP
attribute. This is especially dangerous when combined with the
registration app.

To prevent this check for new lines in the parameters and throw an
exception if one is found. Other illegal characters should be handled
by ldap_add.

Fix issue #145.

Signed-off-by: Alban Bedel <albeu@free.fr>